Fix logic for performing asycn exports
This changes the logic for deciding when to do sync/async exports. Now
we only perform the sync export if the number of visible rows is equal
to the total amount of available transactions.

In every other case, we go async. For large exports ( > 1000 ) we prompt
the merchant, notifying them that they can get a smaller export by
filtering.